refactor: Edit createDt code
