+++ src/main/java/com/neoga/boltauction/memberstore/review/domain/Review.java
@@ -5,6 +5,7 @@ import com.neoga.boltauction.memberstore.store.domain.Store;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
+import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
@@ -24,6 +25,7 @@ public class Review {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "store_id")
     private Store store;
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
     @CreationTimestamp
     @Column(updatable = false)
     private LocalDateTime createDt;
